% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.club.R
\name{print.fun.link}
\alias{print.fun.link}
\title{Prints function corresponding to the function object}
\usage{
\method{print}{fun.link}(x)
}
\value{
\code{fun.link} invisibly
}
\description{
Prints function corresponding to the function object
}
\author{
Vladislav BALAGURA \href{mailto:balagura@cern.ch}{balagura@cern.ch}
}
