% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.club.R
\name{[.fun.club}
\alias{[.fun.club}
\title{Prints function object}
\usage{
\method{[}{fun.club}(
  x,
  link = NULL,
  character.only = FALSE,
  as = c("data.table", "detailed.data.table")
)
}
\description{
Prints function object
}
\author{
Vladislav BALAGURA \href{mailto:balagura@cern.ch}{balagura@cern.ch}
}
