% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.club.R
\name{[.fun.link}
\alias{[.fun.link}
\title{Generates object}
\usage{
\method{[}{fun.link}(x, ...)
}
\description{
\code{...} arguments are forwarded to the function corresponding to
this \code{fun.link} unchanged. For that, the \code{sys.call()} expression is
modified by removing \code{x} first argument and by changing the function
name to \code{generate} of \code{fun.club} class (which will in turn forward
\code{...} to the corresponding function), and then evaluated.

There is a special case when the function corresponding to the \code{fun.link}
also has an \code{x} argument. Then, \code{x=something} given in \code{...} overwrites \code{x}
(originally bounded to the \code{fun.link} object). The code below recovers it,
however, as it is still saved as a first argument in the expression
returned by \code{sys.call()}. (Eg. for \code{a[x=1]} or, equivalently,
\verb{[.fun.link}('a', x=1) \code{sys.call()} returns arguments \verb{('a', x=1)}, and 'a'
can be recovered). So, the function object can use arbitrary names without
restrictions.

Note also, the \code{fun.club} functions in \code{link.methods}, like \code{generate}
below, receive only \code{...} as arguments. This is done on purpose. Eg.
suppose that \code{generate} additionally receives the function object name and
the link number, like in \verb{generate <- function(..., fo, i.link)}. In this
case, if by chance the function object uses the same argument name \code{fo} or
\code{i.link}, and one calls eg. \code{link[a=1, b=2, fo=3]}, then this argument \code{fo = 3} will be in conflict with the argument for the function object name. To
avoid that, all functions receiving arguments through \code{...},
ie. \code{link.methods} and also \code{arg.list(...)}, do not receive anything else
except \code{...}, while the function object and the link number are propagated
externally via \code{link.methods[['fun.object']]} and
\code{link.methods[['i.link']]}.
}
\author{
Vladislav BALAGURA \href{mailto:balagura@cern.ch}{balagura@cern.ch}
}
