% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.club.R
\name{[[<-.fun.link}
\alias{[[<-.fun.link}
\title{Removes generated object from memory but keeps it on disk}
\usage{
\method{[[}{fun.link}(x, ...) <- value
}
\description{
Removes generated object from memory but keeps it on disk
}
\author{
Vladislav BALAGURA \href{mailto:balagura@cern.ch}{balagura@cern.ch}
}
