% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.club.R
\name{[<-.fun.club}
\alias{[<-.fun.club}
\title{Defines generating function in fun.club}
\usage{
\method{[}{fun.club}(x, ..., character.only = FALSE) <- value
}
\description{
Defines generating function in fun.club
}
\author{
Vladislav BALAGURA \href{mailto:balagura@cern.ch}{balagura@cern.ch}
}
