% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.club.R
\name{[<-.fun.link}
\alias{[<-.fun.link}
\title{Deletes \code{fun.link}}
\usage{
\method{[}{fun.link}(x, ...) <- value
}
\description{
Deletes \code{fun.link}
}
\author{
Vladislav BALAGURA \href{mailto:balagura@cern.ch}{balagura@cern.ch}
}
