% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.club.R
\name{unload}
\alias{unload}
\title{Unloads \code{fun.club} from memory in the same way as when quitting R session}
\usage{
unload(fun.club.name)
}
\description{
Unloads \code{fun.club} from memory in the same way as when quitting R session
}
\author{
Vladislav BALAGURA \href{mailto:balagura@cern.ch}{balagura@cern.ch}
}
